First pass at a script to perform terraform plan against all modulues
This is my first Ruby script, so it's a bit of a hack and probably easier
ways to do this.

Similar to the puppet emails, the idea here was to create a (weekly?) script
that could check all of our modules to see if there has been any drift in
what terraform thinks is configured vs what is actually set on the VMs.

Things that still need to be worked out:

1) Not sure at how to get it to run terraform init first then run the plan.

2) The email messages do not contain any new lines between servers, so it's
just one big blob

3) The script has to run terraform plan against everything, so it takes
a long time to run, like at around an hour forty-five.

4) Perhaps figure out a way to integrate it into sulreports.