Add option to cert-approve to create pem certs for Windows
By adding 'pem' as the third option in the cert-approve command,
the script will create a pem file that allow for easy import
into Windows.  The pem file is saved to vault under ssl-pem,
and to be use in role::windows::certs::rdp class to install.

You will have to select 'Microsoft IIS' when
uploading to the tools.stanford.edu SSL site.