Create new script that does DNS lookup for ssl cert creation
Information about the script and how it's run is in the comments
at the top of the ssl-cert-host script, including how to setup
python virtual environments.  You don't have to use the same
virtual environment name as the one in the example, but the
gitignore is setup to ignore any directory that starts with 'env_'.
So as long as you name your virutal environment with that, it shouldn't
be picked up by Git.