Update cert-remctl to include swp flag
This new flag is to be used when creating certificates for aliased
hostnames that are a part of the proxy service.  When that flag and
the skip-netdb flag are set, the certificate is requested and
stored in a different location in vault than normal.  Rather than
under the swp hostname, which can't be accessed by any system due
to lookup restrictions, the ssl keys and certs are placed in
puppet/common/swp/<alias shortname>.

For example, if you wanted to request a certificate for the alias
sulfoo, it would look like:

cert-remctl sulfoo --skip-netdb --swp

The skip-netdb flag is needed because we can only request registered
hostnames by default.  In that example, the vault keys would be
written to:

puppet/common/swp/sulfoo/ssl-key
puppet/common/swp/sulfoo/ssl-cert

Since the keys are put under common, then either proxy host can
access it and be able to load up the files via puppet.

Note that those puppet changes will only work on the new Ubuntu
proxy servers, which are still in the prcoess of being rolled out.
So the swp flag shouldn't be used until then.